+++ docs/isaacsim/snippets/sensors/isaacsim_sensors_camera_depth/updating_existing_assets_to_use_depth_sensors.py
@@ -1,5 +1,14 @@
+import os
+
 from isaacsim.sensors.experimental.rtx import RtxCamera, SingleViewDepthCameraSensor
 
-cam = RtxCamera("/World/depth_sensor")
+usd_path = os.path.join(
+    os.getcwd(),
+    "_example_output_isaacsim.sensors.experimental.rtx",
+    "create_camera_depth_sensor",
+    "example_camera_with_depth_sensor.usd",
+)
+
+cam = RtxCamera.create("/World/depth_sensor", usd_path=usd_path)
 sensor = SingleViewDepthCameraSensor(cam, resolution=(720, 1280), annotators=["depth_sensor_distance"])
 sensor.set_enabled_post_processing(True)